requirements.txt
home-assistant-frontend
ruff
mypy
prek
